requirements.txt
beautifulsoup4==4.5.3
matplotlib==3.0.3
numpy==1.16.2
numpydoc==0.8.0
pandas==0.24.2
pickleshare==0.7.5
requests==2.20.0
requests-toolbelt==0.9.1
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
sklearn==0.0
SQLAlchemy==1.3.1
SQLAlchemy-Continuum==1.3.9
SQLAlchemy-Utils==0.33.11
tqdm==4.31.1
urllib3==1.24.2
adafruit-circuitpython-neopixel==3.4.0
rpi-ws281x==1.0.0
mido==1.2.9
rtmidi==2.3.4
pyaudio==0.2.11
pyalsaaudio==0.8.4
pyalsa==1.1
